Save token in local storage and setting header for future requests.
