+++ package.json
@@ -10,6 +10,7 @@
     "lint": "eslint src --ext .js,.ts,.jsx,.tsx"
   },
   "dependencies": {
+    "@react-native-community/async-storage": "^1.9.0",
     "@react-native-community/masked-view": "^0.1.8",
     "@react-navigation/bottom-tabs": "^5.2.7",
     "@react-navigation/native": "^5.1.5",
